break up bloated es_write_router so that 'admin' and 'bulk' endpoints are in separate modules
